fix: fix memo page's bug - 메모가 아닌 부분을 클릭할 때 오류 발생하는 것 해결
