+++ css/memo-style.css
@@ -34,7 +34,6 @@
   background-color: aliceblue;
   border-radius: 8px;
   padding: 1px 2px;
-  cursor: context-menu;
 }
 
 .item__divider {
@@ -47,6 +46,7 @@
 .item {
   background-color: transparent;
   margin: 8px;
+  cursor: context-menu;
 }
 
 .item-overflow {
