+++ pages/photo.hbs
@@ -1,1 +1,28 @@
-<h1>Photo</h1>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Photo</title>
+  </head>
+  <body>
+    <section class="memo-header">
+      <button class="back" onclick="location.href = '#home'">&lt; BACK</button>
+    </section>
+    {{! <section class="photos"> }}
+    <ul class="photos">
+      <li class="photo__row"><div class="photo"><img src="" alt="" /></div></li>
+      <li class="photo__row"><div class="photo"></div></li>
+      <li class="photo__row"><div class="photo"></div></li>
+      <li class="photo__row"><div class="photo"></div></li>
+      <li class="photo__row"><div class="photo"></div></li>
+      <li class="photo__row"><div class="photo"></div></li>
+      <li class="photo__row"><div class="photo"></div></li>
+      <li class="photo__row"><div class="photo"></div></li>
+      <li class="photo__row"><div class="photo"></div></li>
+      <li class="photo__row"><div class="photo"></div></li>
+    </ul>
+
+    {{! </section> }}
+  </body>
+</html>
